+++ spec/acceptance/groups_spec.rb
@@ -379,7 +379,7 @@ resource "Groups" do
       expect(json["related"].first)
         .to eq({
                  "userId" => group.related_users.first.id,
-                 "state" => group.memberships.first.state,
+                 "state" => group.memberships.first.humanize,
                  "deletedAt" => group.memberships.first.deleted_at
                })
     end